update tests and bugfixes
