+++ sbpy/photometry/iau.py
@@ -8,14 +8,85 @@ created on July, 2022
 """
 
 import warnings
+from collections import OrderedDict
 import numpy as np
 from scipy.integrate import quad
 from astropy.modeling import Parameter
+import astropy.units as u
 from .core import DiskIntegratedPhaseFunc, NonmonotonicPhaseFunctionWarning
 from ..bib import cite
 
 
 __all__ = ['HG', 'HG12BaseClass', 'HG12', 'HG1G2', 'HG12_Pen16']
+
+
+class _spline(object):
+
+    """Cubic spline class
+
+    Spline function is defined by function values at nodes and the first
+    derivatives at both ends.  Outside the range of nodes, the extrapolations
+    are linear based on the first derivatives at the corresponding ends.
+    """
+
+    def __init__(self, x, y, dy):
+        """
+        Spline initialization
+
+        Parameters
+        ----------
+        x, y : array_like float
+            The (x, y) values at nodes that defines the spline
+        dy : array_like float with two elements
+            The first derivatives of the left and right ends of the nodes
+        """
+        from numpy.linalg import solve
+        from numpy.polynomial.polynomial import Polynomial
+        self.x = np.asarray(x)
+        self.y = np.asarray(y)
+        self.dy = np.asarray(dy)
+        n = len(self.y)
+        h = self.x[1:]-self.x[:-1]
+        r = (self.y[1:]-self.y[:-1])/(self.x[1:]-self.x[:-1])
+        B = np.zeros((n-2, n))
+        for i in range(n-2):
+            k = i+1
+            B[i, i:i+3] = [h[k], 2*(h[k-1]+h[k]), h[k-1]]
+        C = np.empty((n-2, 1))
+        for i in range(n-2):
+            k = i+1
+            C[i] = 3*(r[k-1]*h[k]+r[k]*h[k-1])
+        C[0] = C[0]-self.dy[0]*B[0, 0]
+        C[-1] = C[-1]-self.dy[1]*B[-1, -1]
+        B = B[:, 1:n-1]
+        dys = solve(B, C)
+        dys = np.array(
+            [self.dy[0]] + [tmp for tmp in dys.flatten()] + [self.dy[1]])
+        A0 = self.y[:-1]
+        A1 = dys[:-1]
+        A2 = (3*r-2*dys[:-1]-dys[1:])/h
+        A3 = (-2*r+dys[:-1]+dys[1:])/h**2
+        self.coef = np.array([A0, A1, A2, A3]).T
+        self.polys = [Polynomial(c) for c in self.coef]
+        self.polys.insert(0, Polynomial(
+            [self.y[0]-self.x[0]*self.dy[0], self.dy[0]]))
+        self.polys.append(Polynomial(
+            [self.y[-1]-self.x[-1]*self.dy[-1], self.dy[-1]]))
+
+    def __call__(self, x):
+        x = np.asarray(x)
+        out = np.zeros_like(x)
+        idx = x < self.x[0]
+        if idx.any():
+            out[idx] = self.polys[0](x[idx])
+        for i in range(len(self.x)-1):
+            idx = (self.x[i] <= x) & (x < self.x[i+1])
+            if idx.any():
+                out[idx] = self.polys[i+1](x[idx]-self.x[i])
+        idx = (x >= self.x[-1])
+        if idx.any():
+            out[idx] = self.polys[-1](x[idx])
+        return out
 
 
 class HG(DiskIntegratedPhaseFunc):
